+++ src/js/background.js
@@ -183,8 +183,8 @@ const µBlock = {  // jshint ignore:line
 
     // Read-only
     systemSettings: {
-        compiledMagic: 60,  // Increase when compiled format changes
-        selfieMagic: 60,    // Increase when selfie format changes
+        compiledMagic: 61,  // Increase when compiled format changes
+        selfieMagic: 61,    // Increase when selfie format changes
     },
 
     // https://github.com/uBlockOrigin/uBlock-issues/issues/759#issuecomment-546654501